Master Agent/requirements.txt
crewai
python-dotenv
google-genai
pydantic
pdfplumber
pandas
reportlab